Fixed other various issues.
+ Fixed formatting (my IDE has two-space tab sizes).
+ Fixed syntax (forgot a comma ._.).
+ Fixed other text wording.
+ Added a few more lines ( ͡° ͜ʖ ͡°).
- Removed repetitive entry.

Hopefully there aren't any more issues. Hopefully. The reason why the whole file looked changed is because I was editing the formatting to all match each other before, and the easiest thing to do was copy and paste into and out of my IDE. I probably should've set it to JSON though, would've caught all of the errors (except for the formatting). That was on me.